:white_check_mark: finished ex00

+++ L.cpp/cpp_05/ex00/inc/Bureaucrat.hpp
@@ -35,15 +35,15 @@ class Bureaucrat
   private:
 	const std::string _name;
 	size_t _grade;
-	size_t setGrade(int grade);
+	size_t setGrade(size_t grade);
 
   public:
 	// Constructors
 	Bureaucrat();
 	Bureaucrat(Bureaucrat const &src);
-	Bureaucrat(int grade);
+	Bureaucrat(size_t grade);
 	Bureaucrat(const std::string name);
-	Bureaucrat(const std::string name, int grade);
+	Bureaucrat(const std::string name, size_t grade);
 
 	// Overloaded Operators
 	Bureaucrat &operator=(Bureaucrat const &src);
